split react.ts and no export wild
